+++ sdk/ai/azopenai/build.go
@@ -3,7 +3,6 @@
 //go:build go1.18
 // +build go1.18
 
-//go:generate pwsh ./genopenapi3.ps1
 //go:generate autorest  ./autorest.md
 //go:generate go mod tidy
 //go:generate goimports -w .