+++ sdk/ai/azopenai/custom_client_test.go
@@ -92,10 +92,10 @@ func TestGetCompletionsStream_OpenAI(t *testing.T) {
 
 func testGetCompletionsStream(t *testing.T, client *azopenai.Client, tv testVars) {
 	body := azopenai.CompletionsOptions{
-		Prompt:      []string{"What is Azure OpenAI?"},
-		MaxTokens:   to.Ptr(int32(2048)),
-		Temperature: to.Ptr(float32(0.0)),
-		Deployment:  tv.Completions,
+		Prompt:         []string{"What is Azure OpenAI?"},
+		MaxTokens:      to.Ptr(int32(2048)),
+		Temperature:    to.Ptr(float32(0.0)),
+		DeploymentName: &tv.Completions,
 	}
 
 	response, err := client.GetCompletionsStream(context.TODO(), body, nil)
@@ -120,8 +120,8 @@ func testGetCompletionsStream(t *testing.T, client *azopenai.Client, tv testVars
 		}
 
 		if completion.PromptFilterResults != nil {
-			require.Equal(t, []azopenai.PromptFilterResult{
-				{PromptIndex: to.Ptr[int32](0), ContentFilterResults: (*azopenai.PromptFilterResultContentFilterResults)(safeContentFilter)},
+			require.Equal(t, []azopenai.ContentFilterResultsForPrompt{
+				{PromptIndex: to.Ptr[int32](0), ContentFilterResults: safeContentFilterResultDetailsForPrompt},
 			}, completion.PromptFilterResults)
 		}
 
@@ -154,10 +154,10 @@ func TestClient_GetCompletions_Error(t *testing.T) {
 
 	doTest := func(t *testing.T, client *azopenai.Client, model string) {
 		streamResp, err := client.GetCompletionsStream(context.Background(), azopenai.CompletionsOptions{
-			Prompt:      []string{"What is Azure OpenAI?"},
-			MaxTokens:   to.Ptr(int32(2048 - 127)),
-			Temperature: to.Ptr(float32(0.0)),
-			Deployment:  model,
+			Prompt:         []string{"What is Azure OpenAI?"},
+			MaxTokens:      to.Ptr(int32(2048 - 127)),
+			Temperature:    to.Ptr(float32(0.0)),
+			DeploymentName: &model,
 		}, nil)
 		require.Empty(t, streamResp)
 		assertResponseIsError(t, err)